+++ tests/test_init.py
@@ -3,5 +3,5 @@ from src import __version__, __author__
 
 def test_package_metadata_present_and_values():
     """Check that __version__ and __author__ are present and have expected values."""
-    assert isinstance(__version__, str) and __version__ == "0.3.15"
+    assert isinstance(__version__, str) and __version__ == "0.3.16"
     assert isinstance(__author__, str) and __author__ == "Kristoffer Avaldsnes Gilje"