+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.16] - 2025-10-26
+
 ### Added
 
 - `-l` / `--files-with-matches` flag: print only filenames of files containing matches.